+++ src/Microsoft.ApplicationInsights.AspNetCore/Extensions/ApplicationInsightsWebHostBuilderExtensions.cs
@@ -3,6 +3,9 @@
     using Microsoft.ApplicationInsights.AspNetCore.Extensions;
     using Microsoft.Extensions.DependencyInjection;
     using Microsoft.Extensions.DependencyInjection.Extensions;
+#if NETSTANDARD2_0
+    using Microsoft.Extensions.Hosting;
+#endif
     using Microsoft.Extensions.Options;
 
     /// <summary>
@@ -25,6 +28,23 @@
             return webHostBuilder;
         }
 
+#if NETSTANDARD2_0
+        /// <summary>
+        /// Configures <see cref="IHostBuilder"/> to use Application Insights services.
+        /// </summary>
+        /// <param name="hostBuilder">The <see cref="IHostBuilder"/> instance.</param>
+        /// <returns>The <see cref="IHostBuilder"/>.</returns>
+        public static IHostBuilder UseApplicationInsights(this IHostBuilder hostBuilder)
+        {
+            hostBuilder.ConfigureServices(collection =>
+            {
+                collection.AddApplicationInsightsTelemetry();
+            });
+
+            return hostBuilder;
+        }
+#endif
+
         /// <summary>
         /// Configures <see cref="IWebHostBuilder"/> to use Application Insights services.
         /// </summary>
@@ -36,5 +56,19 @@
             webHostBuilder.ConfigureServices(collection => collection.AddApplicationInsightsTelemetry(instrumentationKey));
             return webHostBuilder;
         }
+
+#if NETSTANDARD2_0
+        /// <summary>
+        /// Configures <see cref="IHostBuilder"/> to use Application Insights services.
+        /// </summary>
+        /// <param name="hostBuilder">The <see cref="IHostBuilder"/> instance.</param>
+        /// <param name="instrumentationKey">Instrumentation key to use for telemetry.</param>
+        /// <returns>The <see cref="IHostBuilder"/>.</returns>
+        public static IHostBuilder UseApplicationInsights(this IHostBuilder hostBuilder, string instrumentationKey)
+        {
+            hostBuilder.ConfigureServices(collection => collection.AddApplicationInsightsTelemetry(instrumentationKey));
+            return hostBuilder;
+        }
+#endif
     }
 }