+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 2.8.0-beta1
 - [Fix: Add `IJavaScriptSnippet` service interface and update the `IServiceCollection` extension to register it for `JavaScriptSnippet`.](https://github.com/microsoft/ApplicationInsights-aspnetcore/issues/890)
 - [Fix - ApplicationInsights StartupFilter should not swallow exceptions from downstream ApplicationBuilder.](https://github.com/microsoft/ApplicationInsights-aspnetcore/issues/897)
+- [Add UseApplicationInsights() Extension methods on IHostBuilder](https://github.com/microsoft/ApplicationInsights-aspnetcore/pull/908)
 
 ## Version 2.7.0
 - Updated Web/Base SDK version dependency to 2.10.0